requirements.txt
gradio
openai
python-dotenv
pdfplumber
python-docx
numpy
pandas
scikit-learn
nltk
regex
pydantic
PyMuPDF
re